Added some notes

+++ timing/timer-good.c
@@ -117,7 +117,7 @@ int main(int argc, char** argv) {
      printf("It took a total of %d sec. and %d nanosecs. \n",diff1.tv_sec,(double)diff1.tv_nsec);
 
      */
-    printf("It took a total of %d sec. and %d nanosecs. \n",diff1.tv_sec,(double)diff1.tv_nsec);
+    printf("It took a total of %d sec. and %d nanosecs. \n",diff1.tv_sec,diff1.tv_nsec);
     
     diff_timespec(&start,&end);
